Speed up direct method calls.
For this we avoid creating a bound method object. Then CallAST detects that the
AST it's called on is an AttrAST and the evaluated object implements UL4GetAttr,
it forwards calling the attribute to UL4GetAttr.callAttrUL4(). The default
implementation of that method still does the call in two steps (i.e. get the
attribute, then call it), but the method can be overwritten.

However we still need bound methods when the attribute isn't called directly,
but stored to be called later.